WIP: switch encoding unsigned type normalization to Uint* types
Using Java native types for unsigned values is wasteful especially
in case of uint32 and uint64. Furthermore it leads to uncertainty
between in16/uint8, int32/uint16 and int64/uint32.

This is a partial switch, where we retain java.lang types for
signed entities, but switch the unsigned ones.

+++ yang/yang-data-api/src/main/java/org/opendaylight/yangtools/yang/data/api/codec/Uint16Codec.java
@@ -8,20 +8,17 @@
 package org.opendaylight.yangtools.yang.data.api.codec;
 
 import org.opendaylight.yangtools.concepts.Codec;
+import org.opendaylight.yangtools.yang.common.Uint16;
 
 /**
- * FIXME: In Helium release this codec should be changed to
- *        an UnsignedShort (which needs to be defined somewhere).
+ * A codec between Uint16 and some other representation.
  *
- * @author ttkacik
- *
- * @param <T>
- *          Output type of serialization.
+ * @param <T> Output type of serialization.
  */
-public interface Uint16Codec<T> extends Codec<T,Integer> {
+public interface Uint16Codec<T> extends Codec<T, Uint16> {
     @Override
-    T serialize(Integer data);
+    T serialize(Uint16 data);
 
     @Override
-    Integer deserialize(T data);
+    Uint16 deserialize(T data);
 }
