+++ yang/yang-data-impl/src/test/java/org/opendaylight/yangtools/yang/data/impl/codecs/Uint32CodecStringTest.java
@@ -5,7 +5,6 @@
  * terms of the Eclipse Public License v1.0 which accompanies this distribution,
  * and is available at http://www.eclipse.org/legal/epl-v10.html
  */
-
 package org.opendaylight.yangtools.yang.data.impl.codecs;
 
 import static org.junit.Assert.assertEquals;
@@ -13,6 +12,7 @@ import static org.opendaylight.yangtools.yang.data.impl.codecs.TypeDefinitionAwa
 import static org.opendaylight.yangtools.yang.data.impl.codecs.TypeDefinitionAwareCodecTestHelper.getCodec;
 
 import org.junit.Test;
+import org.opendaylight.yangtools.yang.common.Uint32;
 import org.opendaylight.yangtools.yang.data.api.codec.Uint32Codec;
 import org.opendaylight.yangtools.yang.model.util.type.BaseTypes;
 
@@ -28,7 +28,7 @@ public class Uint32CodecStringTest {
     public void testSerialize() {
         Uint32Codec<String> codec = getCodec(BaseTypes.uint32Type(), Uint32Codec.class);
 
-        assertEquals("serialize", "10", codec.serialize(Long.valueOf(10)));
+        assertEquals("serialize", "10", codec.serialize(Uint32.valueOf(10)));
         assertEquals("serialize", "", codec.serialize(null));
     }
